+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/batch/BulkExecutor.java
@@ -174,7 +174,7 @@ public final class BulkExecutor<TContext> implements Disposable {
         this.executorService.setExecuteExistingDelayedTasksAfterShutdownPolicy(false);
         this.executorService.setRemoveOnCancelPolicy(true);
         this.scheduledFutureForFlush = this.executorService.scheduleWithFixedDelay(
-                () -> onFlush(),
+            this::onFlush,
             this.maxMicroBatchIntervalInMs,
             this.maxMicroBatchIntervalInMs,
             TimeUnit.MILLISECONDS);
@@ -380,7 +380,7 @@ public final class BulkExecutor<TContext> implements Disposable {
                         })
                     .groupBy(Pair::getKey, Pair::getValue)
                     .flatMap(
-                            partitionedGroupFluxOfInputOperations -> executePartitionedGroup(partitionedGroupFluxOfInputOperations),
+                        this::executePartitionedGroup,
                         maxConcurrentCosmosPartitions)
                     .subscribeOn(CosmosSchedulers.BULK_EXECUTOR_BOUNDED_ELASTIC)
                     .doOnNext(requestAndResponse -> {
@@ -561,7 +561,7 @@ public final class BulkExecutor<TContext> implements Disposable {
         ServerOperationBatchRequest serverOperationBatchRequest =
             BulkExecutorUtil.createBatchRequest(operations, pkRange);
         if (serverOperationBatchRequest.getBatchPendingOperations().size() > 0) {
-            serverOperationBatchRequest.getBatchPendingOperations().forEach(t -> groupSink.next(t));
+            serverOperationBatchRequest.getBatchPendingOperations().forEach(groupSink::next);
         }
 
         return Flux.just(serverOperationBatchRequest.getBatchRequest())