+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/GatewayAddressCache.java
@@ -109,12 +109,6 @@ public class GatewayAddressCache implements IAddressCache {
     private final ConnectionPolicy connectionPolicy;
     private final boolean replicaAddressValidationEnabled;
     private final Set<Uri.HealthStatus> replicaValidationScopes;
-    private static final Map<String, OpenConnectionsConcurrencySetting> openConnectionsConcurrencySettings = new HashMap<>();
-
-    static {
-        openConnectionsConcurrencySettings.put("AGGRESSIVE", new OpenConnectionsConcurrencySetting(Configs.getCPUCnt() * 10, Configs.getCPUCnt() * 3));
-        openConnectionsConcurrencySettings.put("DEFENSIVE", new OpenConnectionsConcurrencySetting(Configs.getCPUCnt(), Configs.getCPUCnt()));
-    }
 
     public GatewayAddressCache(
         DiagnosticsClientContext clientContext,
@@ -951,9 +945,6 @@ public class GatewayAddressCache implements IAddressCache {
         checkNotNull(collection, "Argument 'collection' should not be null");
         checkNotNull(partitionKeyRangeIdentities, "Argument 'partitionKeyRangeIdentities' should not be null");
 
-        OpenConnectionsConcurrencySetting connectionsConcurrencySetting = openConnectionsConcurrencySettings
-                .getOrDefault(hint, openConnectionsConcurrencySettings.get("AGGRESSIVE"));
-
         if (logger.isDebugEnabled()) {
             logger.debug(
                     "openConnectionsAndInitCaches collection: {}, partitionKeyRangeIdentities: {}",
@@ -1012,7 +1003,7 @@ public class GatewayAddressCache implements IAddressCache {
                                     addresses.addAll(Arrays.stream(pair.getRight()).toList());
                                 }
                                 return Mono.just(addressInfo);
-                            })
+                            }, Configs.getCPUCnt() * 10, Configs.getCPUCnt() * 3)
                             .flatMap(partitionKeyRangeIdentityPair -> {
                                 AddressInformation[] addressInfosForPkr = partitionKeyRangeIdentityPair.getRight();
                                 return Flux.fromArray(addressInfosForPkr);
@@ -1022,12 +1013,10 @@ public class GatewayAddressCache implements IAddressCache {
 
     }
 
-    public Flux<OpenConnectionResponse> openConnections(
+    public Flux<Void> openConnections(
             AddressInformation address,
             DocumentCollection documentCollection,
-            OpenConnectionAggressivenessHint hint,
-            int connectionsPerEndpointCount,
-            boolean isBackgroundFlow
+            int connectionsPerEndpointCount
     ) {
         if (this.openConnectionsHandler != null) {
 
@@ -1133,16 +1122,6 @@ public class GatewayAddressCache implements IAddressCache {
                 .compareTo(minDurationBeforeEnforcingCollectionRoutingMapRefresh) >= 0;
 
             return returnValue;
-        }
-    }
-
-    private static final class OpenConnectionsConcurrencySetting {
-        private int concurrency;
-        private int prefetch;
-
-        public OpenConnectionsConcurrencySetting(int concurrency, int prefetch) {
-            this.concurrency = concurrency;
-            this.prefetch = prefetch;
         }
     }
 }