+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/ProactiveOpenConnectionsProcessor.java
@@ -13,7 +13,6 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.core.publisher.ParallelFlux;
 import reactor.core.publisher.Sinks;
-import reactor.core.scheduler.Schedulers;
 
 import java.io.Closeable;
 import java.io.IOException;
@@ -21,37 +20,33 @@ import java.net.URI;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
 public final class ProactiveOpenConnectionsProcessor implements Closeable {
     private Sinks.Many<OpenConnectionOperation> openConnectionsTaskSink;
     private Sinks.Many<OpenConnectionOperation> openConnectionsTaskSinkBackUp;
-    private ConcurrentHashMap<String, Integer> tasksInSink;
+    private final ConcurrentHashMap<String, Integer> endpointToMinConnections;
     private static final int MaxConnectionsToOpenAcrossAllEndpoints = 50_000;
     private final AtomicInteger totalEstablishedConnections;
     private final AtomicReference<OpenConnectionAggressivenessHint> aggressivenessHint;
-    private final ScheduledThreadPoolExecutor scheduledThreadPoolExecutor;
-    private ParallelFlux<OpenConnectionResponse> parallelFlux;
 
     public ProactiveOpenConnectionsProcessor() {
         this.openConnectionsTaskSink = Sinks.many().multicast().onBackpressureBuffer();
         this.openConnectionsTaskSinkBackUp = Sinks.many().multicast().onBackpressureBuffer();
         this.aggressivenessHint = new AtomicReference<>(OpenConnectionAggressivenessHint.AGGRESSIVE);
         this.totalEstablishedConnections = new AtomicInteger(0);
-        this.tasksInSink = new ConcurrentHashMap<>();
-        this.scheduledThreadPoolExecutor = new ScheduledThreadPoolExecutor(Configs.getCPUCnt());
+        this.endpointToMinConnections = new ConcurrentHashMap<>();
     }
 
     public synchronized void submitOpenConnectionsTask(OpenConnectionOperation openConnectionOperation) {
 
         String addressUriAsString = openConnectionOperation.getAddressUri().getURIAsString();
 
-        boolean isEndpointTaskInSink = tasksInSink.get(addressUriAsString) != null;
+        boolean isEndpointTaskInSink = endpointToMinConnections.get(addressUriAsString) != null;
 
-        tasksInSink.putIfAbsent(addressUriAsString, openConnectionOperation.getMinConnectionsRequiredForEndpoint());
-        tasksInSink.computeIfPresent(addressUriAsString, (s, min) -> {
+        endpointToMinConnections.putIfAbsent(addressUriAsString, openConnectionOperation.getMinConnectionsRequiredForEndpoint());
+        endpointToMinConnections.computeIfPresent(addressUriAsString, (s, min) -> {
             openConnectionOperation.setMinConnectionsRequiredForEndpoint(Math.max(openConnectionOperation.getMinConnectionsRequiredForEndpoint(), min));
             return Math.max(openConnectionOperation.getMinConnectionsRequiredForEndpoint(), min);
         });
@@ -67,21 +62,6 @@ public final class ProactiveOpenConnectionsProcessor implements Closeable {
 
     }
 
-    public synchronized void toggleOpenConnectionsAggressiveness() {
-        System.out.println("Toggle executor aggressiveness");
-        if (aggressivenessHint.get() == OpenConnectionAggressivenessHint.AGGRESSIVE) {
-            aggressivenessHint.set(OpenConnectionAggressivenessHint.DEFENSIVE);
-            scheduledThreadPoolExecutor.setCorePoolSize(1);
-        } else {
-            aggressivenessHint.set(OpenConnectionAggressivenessHint.AGGRESSIVE);
-            scheduledThreadPoolExecutor.setCorePoolSize(Configs.getCPUCnt());
-        }
-    }
-
-    public ParallelFlux<OpenConnectionResponse> getFlux() {
-        return parallelFlux;
-    }
-
     public ParallelFlux<OpenConnectionResponse> getOpenConnectionsPublisherFromOpenConnectionOperation() {
         // steady-state concurrency
         int concurrency = Configs.getCPUCnt();
@@ -93,10 +73,10 @@ public final class ProactiveOpenConnectionsProcessor implements Closeable {
 
         return Flux.from(openConnectionsTaskSink.asFlux())
                 .parallel(concurrency)
-                .runOn(CosmosSchedulers.OPEN_CONNECTIONS_BOUNDED_ELASTIC).log("Line 90")
+                .runOn(CosmosSchedulers.OPEN_CONNECTIONS_BOUNDED_ELASTIC)
                 .flatMap(openConnectionOperation -> {
                     if (totalEstablishedConnections.get() < MaxConnectionsToOpenAcrossAllEndpoints) {
-                        tasksInSink.remove(openConnectionOperation.getAddressUri().getURIAsString());
+                        endpointToMinConnections.remove(openConnectionOperation.getAddressUri().getURIAsString());
                         IOpenConnectionsHandler openConnectionsHandler = openConnectionOperation.getOpenConnectionsHandler();
                         Uri addressUri = openConnectionOperation.getAddressUri();
                         URI serviceEndpoint = openConnectionOperation.getServiceEndpoint();
@@ -119,7 +99,7 @@ public final class ProactiveOpenConnectionsProcessor implements Closeable {
 
                     if (openConnectionResponse.isConnected()) {
                         totalEstablishedConnections.incrementAndGet();
-                        return Mono.empty();
+                        return Mono.just(openConnectionResponse);
                     }
 
                     return openConnectionOperation
@@ -129,14 +109,20 @@ public final class ProactiveOpenConnectionsProcessor implements Closeable {
                                 if (shouldRetryResult.shouldRetry) {
                                     return enqueueOpenConnectionOpsForRetry(openConnectionOperation, shouldRetryResult);
                                 }
-                                return Mono.empty();
+                                return Mono.just(openConnectionResponse);
                             });
                 });
     }
 
+    public void reinstantiatePublisherAndSubscribe() {
+        this.toggleOpenConnectionsAggressiveness();
+        this.instantiateOpenConnectionsPublisher();
+        this.getOpenConnectionsPublisherFromOpenConnectionOperation().subscribe();
+    }
+
     private Mono<OpenConnectionResponse> enqueueOpenConnectionOpsForRetry(
-        OpenConnectionOperation op,
-        ShouldRetryResult retryResult
+            OpenConnectionOperation op,
+            ShouldRetryResult retryResult
     ) {
         if (retryResult.backOffTime == Duration.ZERO || retryResult.backOffTime == null) {
             this.submitOpenConnectionsTask(op);
@@ -157,5 +143,13 @@ public final class ProactiveOpenConnectionsProcessor implements Closeable {
     public synchronized void instantiateOpenConnectionsPublisher() {
         openConnectionsTaskSink = openConnectionsTaskSinkBackUp;
         openConnectionsTaskSinkBackUp = Sinks.many().multicast().onBackpressureBuffer();
+    }
+
+    public synchronized void toggleOpenConnectionsAggressiveness() {
+        if (aggressivenessHint.get() == OpenConnectionAggressivenessHint.AGGRESSIVE) {
+            aggressivenessHint.set(OpenConnectionAggressivenessHint.DEFENSIVE);
+        } else {
+            aggressivenessHint.set(OpenConnectionAggressivenessHint.AGGRESSIVE);
+        }
     }
 }