+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdOpenConnectionsHandler.java
@@ -11,7 +11,6 @@ import com.azure.cosmos.implementation.ResourceType;
 import com.azure.cosmos.implementation.RxDocumentServiceRequest;
 import com.azure.cosmos.implementation.apachecommons.lang.StringUtils;
 import com.azure.cosmos.implementation.directconnectivity.Uri;
-import com.azure.cosmos.models.OpenConnectionAggressivenessHint;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Flux;
@@ -19,18 +18,14 @@ import reactor.core.publisher.Mono;
 import reactor.core.publisher.SignalType;
 
 import java.net.URI;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
 
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkNotNull;
 
 public class RntbdOpenConnectionsHandler implements IOpenConnectionsHandler {
     private static final Logger logger = LoggerFactory.getLogger(RntbdOpenConnectionsHandler.class);
     private final RntbdEndpoint.Provider endpointProvider;
-    private static final Map<OpenConnectionAggressivenessHint, SemaphoreConfig> openConnectionsSemaphoreRegistry = new HashMap<>();
     private final ProactiveOpenConnectionsProcessor proactiveOpenConnectionsProcessor;
 
     public RntbdOpenConnectionsHandler(RntbdEndpoint.Provider endpointProvider, ProactiveOpenConnectionsProcessor proactiveOpenConnectionsProcessor) {
@@ -89,16 +84,10 @@ public class RntbdOpenConnectionsHandler implements IOpenConnectionsHandler {
                                 OpenConnectionRntbdRequestRecord requestRecord = endpoint.openConnection(requestArgs);
                                 return Mono.defer(() -> Mono.fromFuture(requestRecord).doFinally(signalType -> {
                                     if (signalType.equals(SignalType.CANCEL)) {
-                                        requestRecord.cancel(true);
+                                        requestRecord.cancel(false);
                                     }
                                         }))
                                         .onErrorResume(throwable -> Mono.just(new OpenConnectionResponse(addressUri, false, throwable)))
-                                        .flatMap(openConnectionResponse -> {
-                                            if (!openConnectionResponse.isConnected()) {
-                                                return Mono.error(openConnectionResponse.getException());
-                                            }
-                                            return Mono.just(openConnectionResponse);
-                                        })
                                         .doOnNext(response -> {
 
                                             if (logger.isDebugEnabled()) {
@@ -143,16 +132,4 @@ public class RntbdOpenConnectionsHandler implements IOpenConnectionsHandler {
 
         return openConnectionRequest;
     }
-
-    private static class SemaphoreConfig {
-        private Semaphore semaphore;
-        private int timeoutInMin;
-
-        SemaphoreConfig(int permits, int timeoutInMin) {
-            this.semaphore = new Semaphore(permits);
-            this.timeoutInMin = timeoutInMin;
-        }
-
-    }
-
 }