+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -3,7 +3,9 @@
 ### 4.43.0-beta.1 (Unreleased)
 
 #### Features Added
-Added configurability for minimum connection pool size for all containers through a system property - `COSMOS.MIN_CONNECTION_POOL_SIZE_PER_ENDPOINT` - See [PR 34171](https://github.com/Azure/azure-sdk-for-java/pull/34171).
+* Added configurability for minimum connection pool size for all containers through a system property - `COSMOS.MIN_CONNECTION_POOL_SIZE_PER_ENDPOINT` - See [PR 34171](https://github.com/Azure/azure-sdk-for-java/pull/34171).
+* Added `CosmosClientBuilder:openConnectionsAndInitCaches(CosmosContainerProactiveInitConfig proactiveContainerInitConfig, Duration connectionWarmUpTimeout)` public API to switch between aggressively opening connections
+in a blocking manner to defensively opening connections in a non-blocking manner after `connectionWarmUpTimeout` has elapsed.
 
 #### Breaking Changes